fix: Ensure electrical compatibility for sam pins with "extra" functionality
Previous the pinctrl configuration values were muxed from the atmel hal resulted
in selecting "A" as the peripheral mux functionality when leveraging a
pin's "alternate" or "extra" functionality (which is handled by the EBI).

In some cases this meant that pins which were meant to be used as analog
inputs were driving the pin high whenever the adc is not sampled.

This commit implements logic which checks against the known cases where
this conflict would occur and selects a pio function which is
electrically compatible with the desired "extra"/"alternate" funciton.

Note that this commit also includes a commented out section which
provides an implementation that drives this from the perspective of the
user (in this case the afec driver).

This alternative is more robust with less hardcoded values and is more
explicit/originates from the user which requires the particular pinctrl
config however is not currently possible since the pinctrl config is
const in that context.